+++ laborator9/src/main/java/com/mvc/demo/model/Student.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(name = "first_name")
     private String firstName;
     @Column(name = "last_name")
@@ -33,7 +33,7 @@ public class Student {
 
     public Student() {}
 
-    public Student(Integer id, String firstName, String lastName, Integer age, Integer score){
+    public Student(Long id, String firstName, String lastName, Integer age, Integer score){
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -73,11 +73,11 @@ public class Student {
         this.score = score;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
